break it up into sections, split out float helpers
